Fixed paper errors, better inits
Combed through most of the eqns on page 158 of D'Arco et al EPSR 122 (2014). Found a missing w_g in the last equation, and a sign error in d_iod/dt (should be -cos and not +). Also may have better init terms from some calculations. Disabling damping and enabling voltage feed-forward is listed as a starting pointin III.A of D'Arco et al in IEEE PSCC (2014) and alluded to in the journal version EPES 72 (2015).