Added _extref hooks to take in a v_g and theta_g as external states for upstream power flow
Added a new script called DArcoEPSR122_1bus that runs the DAIIB with a single additional line connecting v_g to the infinite bus, s.t. v_g is no longer fixed at 1 and theta_g is no longer fixed at 0 (though v_inf and theta_inf are, respectively). This requires modifications to current_control and PLL, with the _extref suffix corresponding to the new functions in the DAIIB_extref configuration.

Leverages a version of Jose's pf_eqs in DAIIB_1bus, but using the inverter instead of the machine.

fsolve seems to init even with a rough guess of [1,0] for [v_g, theta_g] and with DAE.